+++ src/Actions/ChooseGrazingField.cs
@@ -2,12 +2,13 @@ using System;
 using System.Linq;
 using Trestlebridge.Interfaces;
 using Trestlebridge.Models;
+using Trestlebridge.Models.Animals;
 
 namespace Trestlebridge.Actions
 {
     public class ChooseGrazingField
     {
-        public static void CollectInput(Farm farm, ISeedProducing seed)
+        public static void CollectInput(Farm farm, IGrazing animal)
         {
             Utils.Clear();
 
@@ -15,16 +16,17 @@ namespace Trestlebridge.Actions
             {
                 Console.WriteLine($"{i + 1}. Grazing Field");
             }
+            // Why are they using a JS style for loop
 
             Console.WriteLine();
 
-            // message to select grazing field
-            Console.WriteLine($"Plant seeds where?");
+            // How can I output the type of animal chosen here?
+            Console.WriteLine($"Place the animal where?");
 
             Console.Write("> ");
             int choice = Int32.Parse(Console.ReadLine());
 
-            farm.GrazingFields[choice-1].AddResource(seed);
+            farm.GrazingFields[choice-1].AddResource(animal);
 
             /*
                 Couldn't get this to work. Can you?
