+++ src/Actions/ChoosePlowedField.cs
@@ -1,0 +1,31 @@
+using System;
+using System.Linq;
+using Trestlebridge.Interfaces;
+using Trestlebridge.Models;
+
+namespace Trestlebridge.Actions
+{
+    public class ChoosePlowedField
+    {
+        public static void CollectInput(Farm farm, ISeedProducing seed)
+        {
+            Utils.Clear();
+
+            for (int i = 0; i < farm.PlowedFields.Count; i++)
+            {
+                Console.WriteLine($"{i + 1}. Plowed Field");
+            }
+            // Why are they using a JS style for loop
+
+            Console.WriteLine();
+
+
+            Console.Write("> ");
+            int choice = Int32.Parse(Console.ReadLine());
+
+            farm.PlowedFields[choice].AddResource(seed);
+
+
+        }
+    }
+}
